Test case improvements

+++ daos/step.js
@@ -17,14 +17,14 @@ module.exports.createStep = async (taskId, stepData) => {
         const updatedTask = await Task.findById(taskId).populate('steps');
         console.log(updatedTask);
 
-        return step;
+        const addedStep = updatedTask.steps.find(addedStep => addedStep._id.toString() === step._id.toString());
+
+        return addedStep;
     } catch (error) {
         console.error(`Error creating step: ${error}`);
         return null;
     }
 };
-
-  
 
 module.exports.getStep = async (id) => {
     return Step.findById(id);
@@ -39,7 +39,19 @@ module.exports.updateStep = async (id, updatedData) => {
 }
 
 module.exports.deleteStep = async (id) => {
+    const stepToDelete = await Step.findById(id);
+    if (!stepToDelete) {
+        console.error(`Step with ID ${id} not found.`);
+        return null;
+    }
+    const task = await Task.findById(stepToDelete.taskId);
+    if (task) {
+        const index = task.steps.indexOf(id);
+        if (index !== -1) {
+            task.steps.splice(index, 1);
+            await task.save();
+        }
+    }
+    
     return Step.findByIdAndRemove(id);
 }
-
-
